Split external theme loading
