+++ doc/architecture/refactor-roadmap.md
@@ -45,6 +45,7 @@ The next refactor should not reduce layering for its own sake. It should turn pa
 - Menu bar state/lifecycle and Nuklear rendering are split across `ui_menubar.c` and `ui_menubar_render.c`.
 - Context menu lifecycle/input handling and Nuklear item rendering are split across `ui_context_menu.c` and `ui_context_menu_render.c`.
 - Theme built-in descriptors and token presets live in `ui_theme_builtin.c`, with private theme-system declarations in `ui_theme_internal.h`.
+- External theme registry state, file/directory reload, reload errors, and directory signatures live in `ui_theme_external.c`.
 - Theme JSON extraction, color parsing, path-derived IDs, override application, and token clamping live in `ui_theme_parse.c`.
 - `render_system_draw()` consumes `RenderSceneDesc`.
 - This refactor round is complete for command routing, workspace API coupling, view-model construction, and low-risk UI composition cleanup.
